// frontend/src/api/modules/system/test/types.ts
export interface SystemTest{
    testId: number ;
    textField: string ;
    numberField: number ;
    dateField: any ;
    richTextField: string ;
    selectField: string ;
    booleanField: string ;
    createTime: any ;
    updateTime: any ;
    textareaField: string ;
    treeField: string ;
}